rename the alloptions fix namespace
